requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

facenet_pytorch==2.5.2
torch==1.8.1+cpu
torchvision==0.9.1+cpu
numpy
Pillow
Flask
joblib
scikit_learn==1.1.1
scikit-image==0.19.3
Werkzeug
Pyrebase4==4.5.0